Removed unsed code / Merged classes
