--- WolvenKit.Modkit/Extensions/CurveExtensions.cs
@@ -1,25 +1,0 @@
-namespace WolvenKit.Modkit.Extensions
-{
-    internal class RedLegacySingleChannelCurveExtensions
-    {
-
-        //    string Elementtype { get; }
-
-        //    public IEditableVariable GetElementInstance(string varName);
-
-        //    public IEnumerable<IREDCurvePoint> GetCurvePoints();
-
-        //    public void SetCurvePoints(IEnumerable<Tuple<double, object>> points);
-
-        //    public ESegmentsLinkType GetLinkType();
-
-        //    public void SetLinkType(ESegmentsLinkType type);
-
-        //    public EInterpolationType GetInterpolationType();
-
-        //    public void SetInterpolationType(EInterpolationType type);
-
-
-    }
-
-}
